Inherit version requirements; Update README
