+++ euphonic_version.py
@@ -5,7 +5,11 @@ import fileinput
 def get_euphonic_version():
     # gets the required euphonic version from `min_requirements.txt` file
     curdir = os.path.dirname(os.path.abspath(__file__))
-    with open(os.path.join(curdir, 'min_requirements.txt')) as minreq:
+    req_file = os.path.join(curdir, 'euphonic_horace', 'min_requirements.txt')
+    if not os.path.isfile(req_file):
+        import update_dependencies
+        update_dependencies.pull_euphonic_horace()
+    with open(req_file, 'r') as minreq:
         verstr = [req for req in minreq if 'euphonic' in req]
     return verstr[0].split('=')[1].strip()
