--- min_requirements.txt
@@ -1,1 +1,0 @@
-euphonic>=0.4.0
